requirements.txt
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
comet_ml==3.49.8
configobj==5.0.9
contourpy==1.3.1
cycler==0.12.1
dulwich==0.22.8
everett==3.1.0
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psutil==7.0.0
py-cpuinfo==9.0.0
Pygments==2.19.1
pyparsing==3.2.3
python-box==6.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
sentry-sdk==2.26.1
setuptools==78.1.0
simplejson==3.20.1
six==1.17.0
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.13.1
tzdata==2025.2
ultralytics==8.3.105
ultralytics-thop==2.0.14
urllib3==2.3.0
wrapt==1.17.2
wurlitzer==3.1.1
pika==1.3.2
python-dotenv==1.0.1
